requirements.txt
flask
tinydb
python-dotenv
google-generativeai
humanize
Werkzeug